separating the report class and the csv generator
